move provider to the lib
